Update tuning tests
