+++ src/torchpme/tuning/pme.py
@@ -5,6 +5,7 @@ from typing import Optional
 import torch
 
 from ..calculators import PMECalculator
+from ..utils import _validate_parameters
 from .tuner import GridSearchTuner, TuningErrorBounds
 
 TWO_PI = 2 * math.pi
@@ -78,6 +79,7 @@ def tune_pme(
     {'interpolation_nodes': 4, 'mesh_spacing': 0.6666666666666666}
 
     """
+    _validate_parameters(charges, cell, positions, exponent)
     min_dimension = float(torch.min(torch.linalg.norm(cell, dim=1)))
     params = [
         {
