+++ tests/utils/test_tuning.py
@@ -10,9 +10,7 @@ from torchpme import (
     P3MCalculator,
     PMECalculator,
 )
-from torchpme.utils.tuning.ewald import EwaldTuner
-from torchpme.utils.tuning.p3m import P3MTuner
-from torchpme.utils.tuning.pme import PMETuner
+from torchpme.tuning import tune_ewald, tune_p3m, tune_pme
 
 sys.path.append(str(Path(__file__).parents[1]))
 from helpers import define_crystal, neighbor_list
@@ -26,15 +24,15 @@ CELL_1 = torch.eye(3, dtype=DTYPE, device=DEVICE)
 
 
 @pytest.mark.parametrize(
-    ("calculator", "tuner", "param_length"),
+    ("calculator", "tune", "param_length"),
     [
-        (EwaldCalculator, EwaldTuner, 1),
-        (PMECalculator, PMETuner, 2),
-        (P3MCalculator, P3MTuner, 2),
+        (EwaldCalculator, tune_ewald, 1),
+        (PMECalculator, tune_pme, 2),
+        (P3MCalculator, tune_p3m, 2),
     ],
 )
 @pytest.mark.parametrize("accuracy", [1e-1, 1e-3, 1e-5])
-def test_parameter_choose(calculator, tuner, param_length, accuracy):
+def test_parameter_choose(calculator, tune, param_length, accuracy):
     """
     Check that the Madelung constants obtained from the Ewald sum calculator matches
     the reference values and that all branches of the from_accuracy method are covered.
@@ -42,15 +40,13 @@ def test_parameter_choose(calculator, tuner, param_length, accuracy):
     # Get input parameters and adjust to account for scaling
     pos, charges, cell, madelung_ref, num_units = define_crystal()
 
-    smearing, params, sr_cutoff = tuner(charges, cell, pos, DEFAULT_CUTOFF).tune(
-        accuracy
-    )
+    smearing, params = tune(charges, cell, pos, DEFAULT_CUTOFF, accuracy=accuracy)
 
     assert len(params) == param_length
 
     # Compute neighbor list
     neighbor_indices, neighbor_distances = neighbor_list(
-        positions=pos, periodic=True, box=cell, cutoff=sr_cutoff
+        positions=pos, periodic=True, box=cell, cutoff=DEFAULT_CUTOFF
     )
 
     # Compute potential and compare against target value using default hypers
@@ -112,32 +108,32 @@ def test_fix_parameters(tune):
     pytest.approx(sr_cutoff, 1.0)'''
 
 
-@pytest.mark.parametrize("tuner", [EwaldTuner, PMETuner, P3MTuner])
-def test_accuracy_error(tuner):
+@pytest.mark.parametrize("tune", [tune_ewald, tune_pme, tune_p3m])
+def test_accuracy_error(tune):
     pos, charges, cell, _, _ = define_crystal()
 
     match = "'foo' is not a float."
     with pytest.raises(ValueError, match=match):
-        tuner(charges, cell, pos, DEFAULT_CUTOFF).tune(accuracy="foo")
+        tune(charges, cell, pos, DEFAULT_CUTOFF, accuracy="foo")
 
 
-@pytest.mark.parametrize("tuner", [EwaldTuner, PMETuner, P3MTuner])
-def test_exponent_not_1_error(tuner):
+@pytest.mark.parametrize("tune", [tune_ewald, tune_pme, tune_p3m])
+def test_exponent_not_1_error(tune):
     pos, charges, cell, _, _ = define_crystal()
 
     match = "Only exponent = 1 is supported"
     with pytest.raises(NotImplementedError, match=match):
-        tuner(charges, cell, pos, DEFAULT_CUTOFF, exponent=2)
+        tune(charges, cell, pos, DEFAULT_CUTOFF, exponent=2)
 
 
-@pytest.mark.parametrize("tuner", [EwaldTuner, PMETuner, P3MTuner])
-def test_invalid_shape_positions(tuner):
+@pytest.mark.parametrize("tune", [tune_ewald, tune_pme, tune_p3m])
+def test_invalid_shape_positions(tune):
     match = (
         r"each `positions` must be a tensor with shape \[n_atoms, 3\], got at least "
         r"one tensor with shape \[4, 5\]"
     )
     with pytest.raises(ValueError, match=match):
-        tuner(
+        tune(
             CHARGES_1,
             CELL_1,
             torch.ones((4, 5), dtype=DTYPE, device=DEVICE),
@@ -146,14 +142,14 @@ def test_invalid_shape_positions(tuner):
 
 
 # Tests for invalid shape, dtype and device of cell
-@pytest.mark.parametrize("tuner", [EwaldTuner, PMETuner, P3MTuner])
-def test_invalid_shape_cell(tuner):
+@pytest.mark.parametrize("tune", [tune_ewald, tune_pme, tune_p3m])
+def test_invalid_shape_cell(tune):
     match = (
         r"each `cell` must be a tensor with shape \[3, 3\], got at least one tensor "
         r"with shape \[2, 2\]"
     )
     with pytest.raises(ValueError, match=match):
-        tuner(
+        tune(
             CHARGES_1,
             torch.ones([2, 2], dtype=DTYPE, device=DEVICE),
             POSITIONS_1,
@@ -161,24 +157,24 @@ def test_invalid_shape_cell(tuner):
         )
 
 
-@pytest.mark.parametrize("tuner", [EwaldTuner, PMETuner, P3MTuner])
-def test_invalid_cell(tuner):
+@pytest.mark.parametrize("tune", [tune_ewald, tune_pme, tune_p3m])
+def test_invalid_cell(tune):
     match = (
         "provided `cell` has a determinant of 0 and therefore is not valid for "
         "periodic calculation"
     )
     with pytest.raises(ValueError, match=match):
-        tuner(CHARGES_1, torch.zeros(3, 3), POSITIONS_1, DEFAULT_CUTOFF)
+        tune(CHARGES_1, torch.zeros(3, 3), POSITIONS_1, DEFAULT_CUTOFF)
 
 
-@pytest.mark.parametrize("tuner", [EwaldTuner, PMETuner, P3MTuner])
-def test_invalid_dtype_cell(tuner):
+@pytest.mark.parametrize("tune", [tune_ewald, tune_pme, tune_p3m])
+def test_invalid_dtype_cell(tune):
     match = (
         r"each `cell` must have the same type torch.float32 as `positions`, "
         r"got at least one tensor of type torch.float64"
     )
     with pytest.raises(ValueError, match=match):
-        tuner(
+        tune(
             CHARGES_1,
             torch.eye(3, dtype=torch.float64, device=DEVICE),
             POSITIONS_1,
@@ -186,14 +182,14 @@ def test_invalid_dtype_cell(tuner):
         )
 
 
-@pytest.mark.parametrize("tuner", [EwaldTuner, PMETuner, P3MTuner])
-def test_invalid_device_cell(tuner):
+@pytest.mark.parametrize("tune", [tune_ewald, tune_pme, tune_p3m])
+def test_invalid_device_cell(tune):
     match = (
         r"each `cell` must be on the same device cpu as `positions`, "
         r"got at least one tensor with device meta"
     )
     with pytest.raises(ValueError, match=match):
-        tuner(
+        tune(
             CHARGES_1,
             torch.eye(3, dtype=DTYPE, device="meta"),
             POSITIONS_1,
